Adding support for base_url prefix
To be able to run behind a reverse proxy under a base_url, I have
added an optional parameter which specifies the URL prefix in use.

Running with 'python zkbrowser.py /zoo_ui' will allow the zkbrowser.py
to be run behind a reverse proxy where the base_url is /zoo_ui.